# Back-end/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Donor Management API",
    "version": "0.0.1-SNAPSHOT",
    "title": "Spring Boot Rest API",
    "contact": {
      "name": "Somnath Musib",
      "url": "medium.com/codefountain",
      "email": "codefountain@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8081",
  "basePath": "/",

  "paths": {
    "/users": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "View all users",
        "operationId": "getUserUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all users",
            "schema": {
              "$ref": "#/definitions/Iterable"
            }
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
    },
    "users/:id":{
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Update a user information",
        "operationId": "UpdateDonorUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated user information",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
    },
    "users/user/:id":{
      "delete": {
        "tags": ["user-controller"],
        "summary": "Deletes specific user with the supplied user id",
        "operationId": "deleteUsingDELETE",
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deletes the specific user"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
    },
    "/auth/signin": {
      "post": {
        "tags": ["user-controller"],
        "summary": "Create a new user",
        "operationId": "createUserUsingPOST",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new User",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["user-controller"],
        "summary": "Retrieve specific user to see if the have signin ",
        "operationId": "getUserUsingPost",
        "produces": ["*/*"],
        "parameters": [
            {
              "in": "body",
              "name": "user",
              "description": "user",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
        "responses": {
          "200": {
            "description": "Successfully logged in ",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "You are not authorized to view the resource"
          },
          "403": {
            "description": "Accessing the resource you were trying to reach is forbidden"
          },
          "404": {
            "description": "The resource you were trying to reach is not found"
          },
          "500": {
            "description": "Application failed to process the request"
          }
        },
        "deprecated": false
      }
     
    }

   
  }
}